+++ skbuild/platform_specifics/windows.py
@@ -3,6 +3,8 @@
 import sys
 import platform
 
+from .abstract import CMakeGenerator
+
 from . import abstract
 
 
@@ -11,7 +13,6 @@ class WindowsPlatform(abstract.CMakePlatform):
 
     def __init__(self):
         super(WindowsPlatform, self).__init__()
-        self.default_generators = ["MinGW Makefiles", ]
         version = sys.version_info
 
         # For Python 2.7 to Python 3.2: VS2008
@@ -19,7 +20,9 @@ class WindowsPlatform(abstract.CMakePlatform):
             (version.major == 2 and version.minor >= 7) or
             (version.major == 3 and version.minor <= 2)
         ):
-            vs_base = "Visual Studio 9 2008"
+            self.default_generators.append(
+                CMakeVisualStudioIDEGenerator("2008")
+            )
 
         # For Python 3.3 to Python 3.4: VS2010
         elif (
@@ -28,20 +31,45 @@ class WindowsPlatform(abstract.CMakePlatform):
                 version.minor <= 4
             )
         ):
-            vs_base = "Visual Studio 10 2010"
+            self.default_generators.append(
+                CMakeVisualStudioIDEGenerator("2010")
+            )
 
         # For Python 3.5 and above: VS2015
         elif version.major == 3 and version.minor >= 5:
-            vs_base = "Visual Studio 14 2015"
+            self.default_generators.append(
+                CMakeVisualStudioIDEGenerator("2015")
+            )
 
         else:
             raise RuntimeError("Only Python >= 2.7 is supported on Windows.")
 
+        self.default_generators.append(
+            CMakeGenerator("MinGW Makefiles")
+        )
+
+
+VS_YEAR_TO_VERSION = {
+    "2008": 9,
+    "2010": 10,
+    "2015": 14
+}
+
+
+class CMakeVisualStudioIDEGenerator(CMakeGenerator):
+    """
+    Represents a Visual Studio CMake generator.
+
+    .. automethod:: __init__
+    """
+    def __init__(self, year):
+        """Instantiate a generator object with its name set to the `Visual
+        Studio` generator associated with the given ``year`` and
+        the current platform (32-bit or 64-bit).
+        """
+        vs_version = VS_YEAR_TO_VERSION[year]
+        vs_base = "Visual Studio %s %s" % (vs_version, year)
         # Python is Win64, build a Win64 module
         if platform.architecture()[0] == "64bit":
             vs_base += " Win64"
-
-        # we're implicitly doing nothing for 32-bit builds.  Their generator
-        # string IDs seem to be just the vs_base.
-
-        self.default_generators.insert(0, vs_base)
+        self._generator_name = vs_base